pyproject.toml
[project]
name = "vectordb-rag-agents-material"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.1.1",
    "crewai>=1.6.1",
    "crewai-tools>=1.6.1",
    "docx2txt>=0.9",
    "ipykernel>=7.1.0",
    "langchain>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-experimental>=0.4.0",
    "langchain-groq>=1.1.0",
    "langchain-openai>=1.1.0",
    "matplotlib>=3.10.7",
    "pypdf>=6.4.0",
    "python-dotenv>=1.2.1",
    "pytube>=15.0.0",
    "sentence-transformers>=5.1.2",
    "unstructured>=0.18.21",
    "wikipedia>=1.4.0",
    "youtube-transcript-api>=1.2.3",
]
